recommendation_parser/requirements.txt
beautifulsoup4~=4.12.2
pyyaml~=6.0.1
